Combine variable declaration rule
